requirements.txt
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
greenlet==1.1.2
idna==3.3
joblib==1.1.0
mypy-extensions==0.4.3
nltk==3.6.5
numpy==1.21.5
pandas==1.3.5
pathspec==0.9.0
platformdirs==2.4.0
pyaml==21.10.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
tomli==1.2.3
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
